Beginning to plan out core functionality
Wrote a large block of notes in index.html spelling out what my app's
core needs to accomplish to hit MVP. To summarize, most functionality
will be carried out with the click of two buttons.

The first is called when a user adds a new movie to the list,
creating the movie in the API and appending the movie to the list.

The second is called when the user decides to remove a movie from their
list. This will remove the item from the API and update the list appropriately.

The first is more straightforward than the second, though there is going
to be more required functionality. The second is modeled for me in the
solution branch of the handlebars lesson, which I plan to reverse engineer
to look at how deletion works.

Next step is conceptualizing js workflow. I created the main.js file to
store the primary front end functionality, but left it empty for now.

With no API, I am well aware that I will eventually hit a point where I
am unable to move forward until I have a fully functioning API. That said,
I hope to do as much conceptual legwork as humanly possible before I need
to switch gears and build the API.