Building skeleton core already FUBAR
Attempting to model core functionality HTML. At simplest, this entails
a button. However, whenever I attempt to create the button, then click
on it, it throws me to the following URL

http://localhost:7165/?submit=Add+new+fave%21

and logs me out. Unsure of why this is the case. Dummied initial code
back into UI in the event that it is the source of the problem. May just
strip down submit to a button, then build out from there.