Prepping for further deployment
With my backend currently up and running, time has come to work on front
end.

purged main file and directed all event handling to be dealt with in movie/events
retaining main file as a code base/ a storage file

Some moderate stylistic changes. Most notably to the font.

There are a lot of things which have been done on this branch which
have yet to be deployed. Consider it wise to do so before moving forward
in building core functionality.