Started working on handlebar manifestation
Built handlebars for index method and integrated them with the front end

I get the strong feeling that there is going to be an ungodly amount of
complexity with the handlebars. May be good for me to take some time to
map out the relationships of all the crud factions once the front end
has completed.