Implemented a jury-rigged delete functionality
Realized that I have delete functionality implemented, but it is done so
with form fields. This means that I could easily slap these forms in the
handlebars, provide an ID reference in the same templates and create an
ugly, but workable method of deletion from within the list.

Next step is tearing up delete HTML and commenting out the method chain
for destroy in movies.

Next step is to look into how I can go about pulling the id of a movie
from the list without needing form fields. This will make delete a one
button feature and will certainly be relevant when I implement my update
functionality.