Update functionality built
Needed to build a new handlebars template in order to display the movie.
This template could prove valuable in providing more specified outputs.

Noticed a weird display bug (though bug feels hyperbolic): any item that
is patched will be placed at the bottom of the list the next time index
is called for whatever reason.When attempting to create some sort of
ranking functionality, this will become problematic.