Same problem as before
Still a fiasco. Core of the problem is that I need to pass both an event
and the ajax results to the success method. As it was written, I didn't
know how to do this; event was an implicit return and I did not know how
to call it explicitly. Trying a new conceptual approach where I pass the
Ajax call through event success as an argument, allowing it to run the
method with both the ajax data it needs and the event it targets.

Also fixed a minor bug where poorly written code in my handlebars script
stopped my update form from resetting.